Move device list demo to dashboard
