+++ src/renderer/containers/OverviewScreen.js
@@ -1,68 +1,58 @@
-import React from 'react';
+/* eslint no-underscore-dangle: ["error", { "allow": ["_id"] }] */
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+
+import AdminApiClient from '../utils/adminApiClient';
+import Overflow from '../components/Overflow';
 import { BarChart, CountsWidget, InterviewWidget, LineChart, PieChart, ServerPanel } from '../components';
+import { countsData, interviewData, barData, pieData, lineData } from './dummy_data';
 
-const countsData = [
-  { name: 'Node count', count: 423 },
-  { name: 'Edge count', count: 256 },
-  { name: 'Interview count', count: 56 },
-];
-
-const interviewData = [
-  { name: 'Duration (minutes)',
-    data: [{ name: 'Mean', count: 23 },
-      { name: 'Min', count: 12 },
-      { name: 'Max', count: 56 }] },
-  { name: 'Node count',
-    data: [{ name: 'Mean', count: 15 },
-      { name: 'Min', count: 2 },
-      { name: 'Max', count: 67 }] },
-  { name: 'Edge count',
-    data: [{ name: 'Mean', count: 12 },
-      { name: 'Min', count: 2 },
-      { name: 'Max', count: 78 }] },
-];
-
-
-const barData = [
-  { name: 'Page A', uv: 4000, pv: 2400, amt: 2400 },
-  { name: 'Page B', uv: 3000, pv: 1398, amt: 2210 },
-  { name: 'Page C', uv: 2000, pv: 9800, amt: 2290 },
-  { name: 'Page D', uv: 2780, pv: 3908, amt: 2000 },
-  { name: 'Page E', uv: 1890, pv: 4800, amt: 2181 },
-  { name: 'Page F', uv: 2390, pv: 3800, amt: 2500 },
-  { name: 'Page G', uv: 3490, pv: 4300, amt: 2100 },
-];
-
-const pieData = [
-  { name: 'Group A', value: 400 },
-  { name: 'Group B', value: 300 },
-  { name: 'Group C', value: 300 },
-  { name: 'Group D', value: 200 },
-  { name: 'Group E', value: 100 },
-  { name: 'Group F', value: 50 },
-  { name: 'Group G', value: 150 }];
-
-// days with no data are represented with null values
-const lineData = [
-  { time: new Date(2017, 6, 24).toLocaleDateString(), value: 40, other: 22 },
-  { time: new Date(2017, 6, 25).toLocaleDateString() },
-  { time: new Date(2017, 6, 26).toLocaleDateString(), value: 75, other: 43 },
-  { time: new Date(2017, 6, 27).toLocaleDateString(), value: 32, other: 45 },
-  { time: new Date(2017, 6, 28).toLocaleDateString(), value: 20, other: 67 },
-  { time: new Date(2017, 6, 29).toLocaleDateString() },
-  { time: new Date(2017, 6, 30).toLocaleDateString(), value: 100, other: 56 },
-  { time: new Date(2017, 7, 1).toLocaleDateString(), value: 5, other: 75 },
-  { time: new Date(2017, 7, 2).toLocaleDateString(), other: 61 },
-  { time: new Date(2017, 7, 3).toLocaleDateString(), value: 15, other: 89 }];
-
-
-export default () => (
-  <div className="overview-dashboard">
-    <ServerPanel className="overview-dashboard__panel overview-dashboard__panel--server-stats" />
-    <div className="overview-dashboard__panel"><CountsWidget data={countsData} /></div>
-    <div className="overview-dashboard__panel"><InterviewWidget data={interviewData} /></div>
-    <div className="overview-dashboard__panel"><BarChart data={barData} /></div>
-    <div className="overview-dashboard__panel"><PieChart data={pieData} /></div>
-    <div className="overview-dashboard__panel"><LineChart data={lineData} /></div>
-  </div>
+const DeviceList = ({ devices }) => (
+  <ol>
+    {devices.map(d => <li key={d._id}>ID: {d._id}</li>)}
+  </ol>
 );
+
+DeviceList.propTypes = {
+  devices: PropTypes.array.isRequired,
+};
+
+class OverviewScreen extends Component {
+  constructor(props) {
+    super(props);
+    this.apiClient = new AdminApiClient();
+    this.state = { devices: [] };
+  }
+
+  componentDidMount() {
+    this.getDevices();
+  }
+
+  getDevices() {
+    this.apiClient.get('/devices')
+      .then(resp => resp.devices)
+      .then(devices => this.setState(devices ? { devices } : {}));
+  }
+
+  render() {
+    const devices = this.state.devices;
+    return (
+      <div className="overview-dashboard">
+        <ServerPanel className="overview-dashboard__panel overview-dashboard__panel--server-stats" />
+        <div className="overview-dashboard__panel">
+          <h2>Paired Devices</h2>
+          <Overflow>
+            <DeviceList devices={devices} />
+          </Overflow>
+        </div>
+        <div className="overview-dashboard__panel"><CountsWidget data={countsData} /></div>
+        <div className="overview-dashboard__panel"><InterviewWidget data={interviewData} /></div>
+        <div className="overview-dashboard__panel"><BarChart data={barData} /></div>
+        <div className="overview-dashboard__panel"><PieChart data={pieData} /></div>
+        <div className="overview-dashboard__panel"><LineChart data={lineData} /></div>
+      </div>
+    );
+  }
+}
+
+export default OverviewScreen;
