+++ src/renderer/containers/__tests__/SettingsScreen-test.js
@@ -3,8 +3,6 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import SettingsScreen from '../SettingsScreen';
 
-jest.mock('../../utils/adminApiClient');
-
 describe('<SettingsScreen />', () => {
   it('should render', () => {
     const subject = shallow((
