+++ src/renderer/styles/components/_all.scss
@@ -3,6 +3,7 @@
 @import './header';
 @import './interview-widget';
 @import './modal';
+@import './overflow';
 @import './pairing_pin';
 @import './pairing_prompt';
 @import './server_panel';
